# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API documentation",
        "description": "API for Waitlists",
        "contact": {
            "url": "https://github.com/LinijaVP"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Api server"
        }
    ],
    "paths": {
        "/customer": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get list of customers",
                "description": "Returns list of customers",
                "operationId": "getCustomers",
                "parameters": [
                    {
                        "name": "name[eq]",
                        "in": "query",
                        "description": "Optional parameter to search by name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "country[eq]",
                        "in": "query",
                        "description": "Optional parameter to search by country",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type[eq]",
                        "in": "query",
                        "description": "Optional parameter to search by country",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "S",
                                "G"
                            ]
                        }
                    },
                    {
                        "name": "showWantlists",
                        "in": "query",
                        "description": "Optional parameter to hide the customers wantlists",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Creates a customer",
                "description": "Create a new customer",
                "operationId": "postCustomers",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCustomerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/customer/{id}": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get the customer with given id",
                "description": "Returns customer with given id",
                "operationId": "getCustomerById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Customer id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "showWantlists",
                        "in": "query",
                        "description": "Optional parameter to hide the customers wantlists",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Customers"
                ],
                "summary": "Edit a customer",
                "description": "Edit a customer",
                "operationId": "editCustomer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Customer id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCustomerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Customers"
                ],
                "summary": "Delete a customer",
                "description": "Delete a customer",
                "operationId": "deleteCustomer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Customer id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Customers"
                ],
                "summary": "Patch a customer",
                "description": "Patch a customer",
                "operationId": "patchCustomer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Customer id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchCustomerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreCustomerRequest": {
                "required": [
                    "name",
                    "type",
                    "email",
                    "city",
                    "country",
                    "budget"
                ],
                "properties": {
                    "name": {
                        "description": "Get the validation rules that apply to the request.",
                        "type": "string",
                        "example": "Name"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "S",
                            "G"
                        ],
                        "example": "S"
                    },
                    "email": {
                        "type": "email",
                        "example": "vp@gmail.com"
                    },
                    "city": {
                        "type": "string",
                        "example": "Portland"
                    },
                    "country": {
                        "type": "string",
                        "example": "Oregon"
                    },
                    "budget": {
                        "type": "integer",
                        "example": 1000
                    }
                },
                "type": "object"
            },
...
            "PatchCustomerRequest": {
                "properties": {
                    "name": {
                        "description": "Get the validation rules that apply to the request.",
                        "type": "string",
                        "example": "Name"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "S",
                            "G"
                        ],
                        "example": "S"
                    },
                    "email": {
                        "type": "email",
                        "example": "vp@gmail.com"
                    },
                    "city": {
                        "type": "string",
                        "example": "Portland"
                    },
                    "country": {
                        "type": "string",
                        "example": "Oregon"
                    },
                    "budget": {
                        "type": "integer",
                        "example": 1000
                    }
                },
                "type": "object"
            },
...
            "CustomerResource": {
                "properties": {
                    "name": {
                        "description": "Transform the resource into an array.",
                        "type": "string",
                        "example": "Name"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "S",
                            "G"
                        ],
                        "example": "S"
                    },
                    "email": {
                        "type": "email",
                        "example": "vp@gmail.com"
                    },
                    "city": {
                        "type": "string",
                        "example": "Portland"
                    },
                    "country": {
                        "type": "string",
                        "example": "Oregon"
                    },
                    "budget": {
                        "type": "integer",
                        "example": 1000
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "name": "sanctum",
                "in": "header",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}